clean up codebase
