--- verl/workers/sharding_manager/fsdp/fsdp_sglang.py
@@ -1,133 +1,0 @@
-# Copyright 2024 Bytedance Ltd. and/or its affiliates
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-import os
-import logging
-import torch
-from torch.distributed.fsdp.fully_sharded_data_parallel import FullyShardedDataParallel as FSDP
-from torch.distributed.fsdp.api import ShardingStrategy, ShardedStateDictConfig, StateDictType, FullStateDictConfig
-from torch.distributed.device_mesh import DeviceMesh
-
-from verl import DataProto
-from verl.utils.torch_functional import (broadcast_dict_tensor, allgather_dict_tensors)
-from verl.utils.debug import log_gpu_memory_usage
-from sglang.srt.entrypoints.verl_engine import VerlEngine
-from ..base import BaseShardingManager
-from verl.third_party.sglang import parallel_state as sglang_ps
-# from vllm.distributed import parallel_state as sglang_ps
-
-logger = logging.getLogger(__file__)
-logger.setLevel(os.getenv('VERL_PPO_LOGGING_LEVEL', 'WARN'))
-
-
-class FSDPSGLangShardingManager(BaseShardingManager):
-
-    def __init__(self,
-                 module: FSDP,
-                 inference_engine: VerlEngine,
-                 model_config,
-                 full_params: bool = False,
-                 device_mesh: DeviceMesh = None):
-        self.module = module
-        self.inference_engine = inference_engine
-        self.model_config = model_config
-        self.device_mesh = device_mesh
-
-        # Full params
-        self.full_params = full_params
-        if full_params:
-            FSDP.set_state_dict_type(self.module,
-                                     state_dict_type=StateDictType.FULL_STATE_DICT,
-                                     state_dict_config=FullStateDictConfig())
-        else:
-            FSDP.set_state_dict_type(self.module,
-                                     state_dict_type=StateDictType.SHARDED_STATE_DICT,
-                                     state_dict_config=ShardedStateDictConfig())
-
-        # Note that torch_random_states may be different on each dp rank
-        self.torch_random_states = torch.cuda.get_rng_state()
-        # get a random rng states
-        if self.device_mesh is not None:
-            gen_dp_rank = self.device_mesh['dp'].get_local_rank()
-            torch.cuda.manual_seed(gen_dp_rank + 1000)  # make sure all tp ranks have the same random states
-            self.gen_random_states = torch.cuda.get_rng_state()
-            torch.cuda.set_rng_state(self.torch_random_states)
-        else:
-            self.gen_random_states = None
-
-    def __enter__(self):
-        log_gpu_memory_usage('Before state_dict() in sharding manager memory', logger=logger)
-        params = self.module.state_dict()
-        log_gpu_memory_usage('After state_dict() in sharding manager memory', logger=logger)
-        # Copy, not share memory
-        load_format = None if self.full_params else 'dtensor'
-
-        self.inference_engine.update_weights_from_tensor([(k, v) for k, v in params.items()], load_format=None)
-        log_gpu_memory_usage('After sync model weights in sharding manager', logger=logger)
-
-        del params
-        torch.cuda.empty_cache()
-        log_gpu_memory_usage('After del state_dict and empty_cache in sharding manager', logger=logger)
-
-        # TODO: offload FSDP model weights
-        # self.module.cpu()
-        # torch.cuda.empty_cache()
-        # if torch.distributed.get_rank() == 0:
-        # print(f'after model to cpu in sharding manager memory allocated: {torch.cuda.memory_allocated() / 1e9}GB, reserved: {torch.cuda.memory_reserved() / 1e9}GB')
-
-        # important: need to manually set the random states of each tp to be identical.
-        if self.device_mesh is not None:
-            self.torch_random_states = torch.cuda.get_rng_state()
-            torch.cuda.set_rng_state(self.gen_random_states)
-
-    def __exit__(self, exc_type, exc_value, traceback):
-        log_gpu_memory_usage('Before SGLang offload in sharding manager', logger=logger)
-        self.inference_engine.release_memory_occupation
-        log_gpu_memory_usage('After SGLang offload in sharding manager', logger=logger)
-
-        # self.module.to('cuda')
-        # if torch.distributed.get_rank() == 0:
-        #     print(f'after actor module to cuda in sharding manager memory allocated: {torch.cuda.memory_allocated() / 1e9}GB, reserved: {torch.cuda.memory_reserved() / 1e9}GB')
-
-        self.module.train()
-
-        # add empty cache after each compute
-        torch.cuda.empty_cache()
-
-        # restore random states
-        if self.device_mesh is not None:
-            self.gen_random_states = torch.cuda.get_rng_state()
-            torch.cuda.set_rng_state(self.torch_random_states)
-
-    def preprocess_data(self, data: DataProto) -> DataProto:
-        # TODO: Current impl doesn't consider FSDP with torch micro-dp
-        data.batch = allgather_dict_tensors(data.batch.contiguous(),
-                                            size=self.device_mesh["infer_tp"].mesh.size()[0],
-                                            group=self.device_mesh["infer_tp"].get_group(),
-                                            dim=0)
-        return data
-
-    def postprocess_data(self, data: DataProto) -> DataProto:
-        # TODO: Current impl doesn't consider FSDP with torch micro-dp
-        global_rank = self.device_mesh.get_rank()
-        tp_rank = self.device_mesh["infer_tp"].get_local_rank()
-        tp_size = self.device_mesh["infer_tp"].mesh.size()[0]
-        src_rank = global_rank // tp_size * tp_size
-        broadcast_dict_tensor(data.batch,
-                              src=src_rank,
-                              group=self.device_mesh["infer_tp"].get_group())
-        if tp_size > 1:
-            local_prompts = data.chunk(chunks=tp_size)
-            data = local_prompts[tp_rank]
-        return data
